Build upcgen.py into a simple but functional bulk barcode generator.
Configure upcgen.py to read a list of barcodes from barcodes.txt and to
generate a set of images and a web page that groups them together.  The
web page is based on a template, which should make it easy to tweak the
style later.

Error-handling should probably be improved later, and maybe allow for
command-line arguments to override some built-in defaults...