Remove old barcode-generating code/data.
Get rid of the old mkbarcode utility; upcgen.py should now be good
enough for us (even if it doesn't have all the features), and with our
own code, we don't need to worry about licenses.

Also get rid of the administrator barcode (no longer used) and the
barcode-layout Word document (we'll try HTML for now, or can use TeX too
if needed).