Add a "barcode" column to the transactions table.
Start tracking the barcodes of items purchased, in addition to storing a
text description.  This should make post-processing of the database
simpler, though it will be a while before there are barcodes for most
purchases.

The barcode column is left NULL for when privacy is enabled, or for
transactions that are not item purchases.