Add support for aggregate purchase tracking.
I've added another table to the database, aggregate_purchases, that
tracks which items have been purchased when, but without user info and
only at the granularity of a day.  This will be used for tracking
aggregate purchase information, and it should be possible to track items
purchased by people with the privacy setting on (since the
low-resolution timestamp should mean there will not be enough
information available to tie the transaction back to the person...unless
there are very few people with privacy enabled...perhaps we should
advertise the option a bit more to provide more cover?).

This will eventually replace the direct tracking of product inventory,
since it will provide slightly more useful information.