Track "Buy Other" items even when privacy is enabled.
Keep track of whether a transaction was the result of entering a
specific item or of selecting "Buy Other", even when privacy is enabled.
Transactions will show up as either "BUY" or "BUY OTHER" when privacy is
turned on.  This will allow the total "Buy Other" amount to be
accurately calculated; this is important since it is the only
transaction that does not update the aggregate purchases table.

At the same time, clean out some references to the old purchase
categories ("Buy Soda"/"Buy Candy"/etc.) since these no longer show up
in the menu.