When preparing an order, always use the most recent inventory data.
Previously, if the inventory was taken the same day as the order
estimate, the previous inventory was used.  This was because the
inventory estimate function was producing output suitable for an
inventory-entry page, where the we do want to display data based on the
previous inventory.  Now, make the behavior selectable.