Rework handling of negative inventories in inventory value estimation.
The old version led to some pathological cases triggered when
non-existent items (according to the estimate) were sold, and resulted
in possibly larger inventory valuations than were correct.  The new
version should be more principled: we always consider the value of zero
or negative number of an item to be zero.