Allow inventory cost estimates to run on just a subset of dates.
Change the inventory summarization code so that it can rewrite the data
for just more recent dates (this will be important if we want to keep
data from old financial periods stable).  However, still do a
computation over a slightly longer time period, so that the estimates
will tend to converge to the same value instead of having a
discontinuity at the start of the period.