Track the number of items in a case in the inventory tracking code.
By storing the case size in the inventory table, we can support changing
the size of a case while still keeping the old inventory records (which
include a count of cases and loose items) valid.

In the take_inventory code, use the value stored in the inventory table
so that when looking at or editing old inventories, the case size
displayed and used is that stored in the table, not the current value in
the bulk_items table.