Begin converting web site to Django 1.0.
Start the process of converting the Chez Bob administrative web
interface to Django 1.0 (the old code is based on Django 0.95.1, which
is what comes with Debian etch).

The changes so far update the database models and should get the
automatic admin interface working.  Other custom code does not yet work
(most of the remaining changes are probably due to the change from float
to decimal as the base type for monetary fields, but there may be other
changes as well).