Implement a top-level query menu.
This presents the available database queries as a series of forms (with
entries for each variable in the query), with a button to execute the
query.

This is for the non-JavaScript version of the query page.  The
JavaScript-enabled version will use an AJAX interface to fill in the
results directly; this is not implemented yet.