Update SQL mainenance scripts.
When pruning old transactions, do not any longer bother to insert a
starting 'INIT' transaction if it has a zero balance.  This means that
old accounts with a zero balance will have no transactions listed, and
only non-zero account balances will get the 'INIT' transaction.
Newly-opened accounts will continue to have an 'INIT' transaction on the
opening date, until that transaction is expired.