Delete the books database backup.
This was obsolete even before this repository was created, and all the
books probably refer to items in the old building that no longer exist.
If anyone does care about the data, it does at least exist in the
revision history, but it's not really important enough to keep the file
around in the current revision.