Create a script for repricing and printing a price list.
We can now produce a web page that lists current prices and the previous
price for an item.  When generating this page, the prices are updated.
We also track historical prices for each item, so that we can find the
previous price.  Running the update script multiple times should be
safe, and should continue to provide good data on the previous price.